# Improve performance of buffered string write