When the string to be written can potentially be fully buffered there is a second call to `GetByteCount` (first one is GetSize). If the converter is not nested inside another converter, the size has already been calculated.
Basing this off BitArrayBitStringConverter that its safe to do.

PgReader.cs change is just to remove the DisplayClass struct creation.

Text.cs test string of 10_000 doesn't fit in the test buffer.

**Before**
| Method | Value                | Mean      | Error    | StdDev   | Op/s         | Gen0   | Allocated |
|------- |--------------------- |----------:|---------:|---------:|-------------:|-------:|----------:|
| Read   | x                    |  25.41 ns | 0.203 ns | 0.190 ns | 39,359,074.1 | 0.0023 |      24 B |
| Write  | x                    |        NA |       NA |       NA |           NA |     NA |        NA |
| Read   | xxxxxxxxxx           |  27.39 ns | 0.370 ns | 0.346 ns | 36,512,150.3 | 0.0046 |      48 B |
| Write  | xxxxxxxxxx           |  20.93 ns | 0.230 ns | 0.204 ns | 47,775,325.5 | 0.0000 |       1 B |
| Read   | xxxx(...)xxxx [100]  |  30.99 ns | 0.196 ns | 0.184 ns | 32,267,999.8 | 0.0214 |     224 B |
| Write  | xxxx(...)xxxx [100]  |  24.14 ns | 0.064 ns | 0.057 ns | 41,425,834.0 | 0.0007 |       8 B |
| Read   | xxxx(...)xxxx [1000] | 107.35 ns | 0.816 ns | 0.637 ns |  9,315,283.3 | 0.1935 |    2024 B |
| Write  | xxxx(...)xxxx [1000] |  74.81 ns | 0.312 ns | 0.261 ns | 13,367,517.5 | 0.0074 |      78 B |

**After**
| Method | Value                | Mean      | Error    | StdDev   | Op/s         | Gen0   | Allocated |
|------- |--------------------- |----------:|---------:|---------:|-------------:|-------:|----------:|
| Read   | x                    |  23.42 ns | 0.095 ns | 0.089 ns | 42,704,740.3 | 0.0023 |      24 B |
| Write  | x                    |        NA |       NA |       NA |           NA |     NA |        NA |
| Read   | xxxxxxxxxx           |  25.19 ns | 0.075 ns | 0.070 ns | 39,701,387.8 | 0.0046 |      48 B |
| Write  | xxxxxxxxxx           |  15.57 ns | 0.039 ns | 0.037 ns | 64,237,944.4 | 0.0000 |         - |
| Read   | xxxx(...)xxxx [100]  |  31.36 ns | 0.174 ns | 0.145 ns | 31,890,144.5 | 0.0214 |     224 B |
| Write  | xxxx(...)xxxx [100]  |  20.31 ns | 0.058 ns | 0.052 ns | 49,229,262.7 | 0.0005 |       5 B |
| Read   | xxxx(...)xxxx [1000] | 107.48 ns | 0.698 ns | 0.653 ns |  9,304,087.4 | 0.1935 |    2024 B |
| Write  | xxxx(...)xxxx [1000] |  57.45 ns | 0.100 ns | 0.094 ns | 17,405,246.9 | 0.0052 |      54 B |