@NinoFloris 2025-10-29
I'm going to close this draft. The idea is sound, but the approach has some problems. If we want to avoid the extra work we should instead return a writeState from GetSize that contains the encoded length, this would work in all cases. The downside being that it adds an alloc per string write.